images moved into /assets/images
