+++ src/features/themes/ThemeSwitcher/index.js
@@ -1,8 +1,8 @@
 import { useDispatch, useSelector } from "react-redux";
 import { selectTheme, setTheme } from "../themeSlice";
 import { SwitcherButton, SwitcherButtonImage, SwitcherCaption } from "./styled";
-import light from "../../../images/lightMode.svg";
-import dark from "../../../images/darkMode.svg";
+import light from "../../../assets/images/lightMode.svg";
+import dark from "../../../assets/images/darkMode.svg";
 
 const ThemeSwitcher = () => {
     const isThemeDark = useSelector(selectTheme);
